requirements.txt
fastapi==0.85.0
pyserial==3.5
uvicorn==0.17.5